Refactor shell package also
